Version 1.3 (#285)
* reinstate collection.delete_many with empty filter for truncate

* deprecation of delete_all (+related test utilities); removal of old core aliased methods

* Deprecation sine die of bulk_write

* changes

* vector/s/ize params deprecation, collection.py

* resiliency of tests (core+idiomatic) to vector not being returned under certain conditions

* improved vectorize_methods settings handling; partial support for hfded and oaiparams

* shorten model root name

* hf ded tweaks to vectorize coll models generation

* adapt vectorize tests to hf ded

* deprecation of vector/ize orderly brought to all Operation classes

* revert deprecation of bulk_insert

* adapt testing to modelName management in HD dedicated

* tests capture and check deprecations for all usages of vector/ize= standalone params

* added HCD and CASSANDRA to the Environment values

* more clarity in string repr of cursors

* use type alias for filter consistently

* callback mechanism to store the sort vector

* expose include_sort_vector option to find

* add prefetch iterator wrappers for the returnvector

* prefetching for include_sort_vector

* revert nullable-model-name changes and cope with disappeared .providerKey suffix in tests

* include_sort_vector and get_sort_vector, extensive test coverage

* add code examples for get_sort_vector in docstring for the two find methods

* add Content-Type header to all api requests

* changelog update

* token providers

* full transition to TokenProvider

* vectorize: support testing of openai project/org ids

* rename attribute token -> token_provider where needed

* docstring about token provider

* token inheritance tests; eq and copy methods of api commander

* changes file

* tests: modernize and lift upward fixtures and paradigmaticize object creation; add test on tokenless client

* astra/nonastra plug-your-db management

* astra/nonastra management fully covers vectorize

* nonAstra admin (i.e. namespace crud) tests

* rework fixture inheritance and naming for alternate keyspace (incl.astra/nonastra)

+++ tests/idiomatic/integration/test_exceptions_async.py
@@ -31,7 +31,7 @@ from astrapy.exceptions import (
 from astrapy.constants import DocumentType
 from astrapy.cursors import AsyncCursor
 
-from ..conftest import AstraDBCredentials
+from ..conftest import DataAPICredentials, IS_ASTRA_DB
 
 
 class TestExceptionsAsync:
@@ -74,21 +74,21 @@ class TestExceptionsAsync:
         assert len(im_result1.inserted_ids) == 6
         assert len(await _alist(acol.find({}))) == 6
 
-        await acol.delete_all()
+        await acol.delete_many({})
         im_result2 = await acol.insert_many(
             ok_docs, ordered=False, chunk_size=2, concurrency=1
         )
         assert len(im_result2.inserted_ids) == 6
         assert len(await _alist(acol.find({}))) == 6
 
-        await acol.delete_all()
+        await acol.delete_many({})
         im_result3 = await acol.insert_many(
             ok_docs, ordered=False, chunk_size=2, concurrency=2
         )
         assert len(im_result3.inserted_ids) == 6
         assert len(await _alist(acol.find({}))) == 6
 
-        await acol.delete_all()
+        await acol.delete_many({})
         with pytest.raises(InsertManyException) as exc:
             await acol.insert_many(dup_docs, ordered=True, chunk_size=2, concurrency=1)
         assert len(exc.value.error_descriptors) == 1
@@ -98,7 +98,7 @@ class TestExceptionsAsync:
         assert len(exc.value.partial_result.raw_results) == 2
         assert {doc["_id"] async for doc in acol.find()} == {"a", "b"}
 
-        await acol.delete_all()
+        await acol.delete_many({})
         with pytest.raises(InsertManyException) as exc:
             await acol.insert_many(dup_docs, ordered=False, chunk_size=2, concurrency=1)
         assert len(exc.value.error_descriptors) == 3
@@ -109,7 +109,7 @@ class TestExceptionsAsync:
         assert len(exc.value.partial_result.raw_results) == 4
         assert {doc["_id"] async for doc in acol.find()} == {"a", "b", "d", "e", "f"}
 
-        await acol.delete_all()
+        await acol.delete_many({})
         with pytest.raises(InsertManyException) as exc:
             im_result3 = await acol.insert_many(
                 dup_docs, ordered=False, chunk_size=2, concurrency=2
@@ -177,7 +177,7 @@ class TestExceptionsAsync:
         acol._astra_db_collection.collection_name += "_hacked"
         acol._astra_db_collection.base_path += "_hacked"
         with pytest.raises(DataAPIResponseException):
-            await acol.delete_all()
+            await acol.delete_many({})
         with pytest.raises(DataAPIResponseException):
             await acol.delete_one({"a": 1})
         with pytest.raises(DataAPIResponseException):
@@ -293,8 +293,9 @@ class TestExceptionsAsync:
         async_database: AsyncDatabase,
     ) -> None:
         f_database = async_database._copy(namespace="nonexisting")
-        with pytest.raises(DataAPIResponseException):
-            await f_database.drop_collection("nonexisting")
+        if IS_ASTRA_DB:
+            with pytest.raises(DataAPIResponseException):
+                await f_database.drop_collection("nonexisting")
         with pytest.raises(DataAPIResponseException):
             await async_database.command(body={"myCommand": {"k": "v"}})
         with pytest.raises(DataAPIResponseException):
@@ -312,9 +313,9 @@ class TestExceptionsAsync:
     async def test_get_database_info_failures_async(
         self,
         async_database: AsyncDatabase,
-        astra_db_credentials_kwargs: AstraDBCredentials,
+        data_api_credentials_kwargs: DataAPICredentials,
     ) -> None:
-        hacked_ns = (astra_db_credentials_kwargs["namespace"] or "") + "_hacked"
+        hacked_ns = (data_api_credentials_kwargs["namespace"] or "") + "_hacked"
         with pytest.raises(DevOpsAPIException):
             async_database._copy(namespace=hacked_ns).info()
